services/mailserver/usermgmt/requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
